Implement Standalone/Development mode
Enabled standalone run with the usage of webpack-dev-server.
Since the application now expects already authenticated user, `npm start` will
prompt the developer for credentials to the engine instance specified in the
ENGINE_URL environment variable.

The same flow is used in the accompanying docker image.

The injection of the `userData` is facilitated by extending the `env` object
(that is used to source the webpack.DefinePlugin) by the entry of
'window.userData' with the complete object containing token & username.

Note: window.userData is also used by the JSP templates to inject such data
from the engine so we maintain consistency and fetchToken() in index.js remains
the single authoritative source of obtaining pre-existing auth token on
application startup.