package.json not processed by autotools
The version in package.json is permanently set to 0.0.0. This is ok,
since content of this repository is not distributed through npm but
rather rpm. rpm version is stored in `configure.ac` (untouched).

This allows to to use `npm run ...` commands right after cloning of the
repository without the need to run autotools.