Stuck popup fix
Problem: Popovers from react-bootstrap library used in FieldHelp
component was not moving when pane were scrolled. It was caused by the
library creating popover div tags at the end of the document i.e.
outside of the scrolled area.

Solution: Following guide [1] the OverlayTrigger component is passed a
'container' property that set element into which the popover is
rendered. Plus css adjustments.

Component FieldHelp changed to class one for it to be referenceable by
`this` keyword.

[1]: https://react-bootstrap.github.io/components.html#popovers-positioned-scrolling