Fix of scrolling of the grid on Webkit browsers
Problem: There was no scrollbar / no possibility to scroll on Webkit
like browsers - reproduced on Chrome, Chrome for Android and iOS Safari.
It was caused by missing height specification of last (third) element of
PageRouter component. Thus value '100%' of 'height' CSS property of its
child resolved to 'auto' instead to its actual height.

From MDN [1]:
If the height of the containing block is not specified explicitly (i.e.,
it depends on content height), and this element is not absolutely
positioned, the value computes to auto.

[1]: https://developer.mozilla.org/en-US/docs/Web/CSS/height