CD removed from New VM dialog
Problem: If a new VM was based on a template with disks then
`createNewVm` saga first sent a request to create a VM which was
immediately locked for the time of disk image copying. Then the saga
fired a request to set a CD image. However the request failed because
the VM was locked.

Solution: CD image can no longer be customized from New VM page, it is
possible only from Edit VM page.