Fix expired token error message to be fully visible
Due to the way the PageRouter is laid out, the token error message
would display beneath and only be partially visible.  This is fixed
twice by setting the z-index on .alert to be higher, and by moving
the tag below the PageRouter in the App's dom.  Both changes will
make sure the alerts are not hidden.

Refactored OvirtApiCheckFailed to use ErrorAlert, allowing future
changes to error alerts to be done in one place.

Resolves issue #395