View and Edit pages will wait for their VM/Pool to be loaded
Issue #574 was caused when a user would enter the app at a view
or edit page with a VM or Pool that would be on a fetch page >1.
The VM/Pool would be requested but often would either fail the
request or would not arrive prior the initial data load flag
being set true. The concurrency issues would leave the user
either on a blank page or redirected to the card view page.

A number of changes, along with the refactoring of the App init,
react-router page rendering, and login saga, have been made to
 eliminate this issue:

 - The single VM fetch REST call can now __follow__ all the data
   paths like the fetch by page does. This reduces the number of
   REST calls required from 6 to 1. It now also supports both
   shallow and deep fetches.

 - The __fetchSingleVm__ saga is upgraded to tag the VM IDs to
   the redux state that it has successfully fetched

 - The __fetchSinglePool__ saga was upgraded in the same way

 - The redux store **state.vms** reducer can now keep track
   of the VM/Pool IDs that are loaded independent of those loaded
   by the paged loader. Actions and constants were added in
   support of the reducer changes.

 - Selectors were added to fetch those 2 lists from the store

 - The __refreshData__ saga was upgraded so it refreshes both
   the individually fetched and page fetched VMs and Pools
   without dropping any of them from the redux store. The removal
   of 'stale' VMs and Pools from the store was moved to this saga
   so it can be done once and preserve ALL VMs and Pools regardless
   of how they were fetched.