Fixed the 'selecting VM details' unit test by initializing a mock store
Moved the __env__ jest config from CLI to package.json properties
so tests can be run in the background and debugged in vscode.

Changes introduced in the previous commits cause the sagas being
tested to need a mock store. Specifically the __selectors__ needed to
be initialized or they throw an exception. The initial state of reducers
used by __selectors__ are now exported to be used in the unit test.