Updated the `refreshData` saga based on PR review comments
 - Calculate the IDs that need to be fetched individually instead of
   tracking them in the redux store

 - Added parallel `fetchSingle*` calls to speed up the single item
   fetch processing in `refreshData` should there be >1 item not
   refreshed by the `fetch*ByCount` functions

 - Removed the `*FetchedById` states, reducers, actions and constants
   since they're no longer needed

 - Adjusted the selectors to return all VM or Pool IDs currently held
   in the store

 - Fixed a latent bug in the UPDATE_VMPOOLS_COUNT reducer in
   __src/reducers/vms.js__

NOTE: This change will cause `refreshData` to singly fetch an VM
or Pool not fetched by the paged/by count functions. That could be
for any reason including, but not limited to, accessing a VM or Pool
that is past the first fetch page via direct URL (this PR's issue),
deletion of the VM or Pool, permission changes to the VM or Pool,
or if the VM or Pool gets pushed past the last page the App is
currently refreshing.