Update `refreshData` saga to fetch single VM/Pool as needed
Previously, refresh data ignored and removed any VM/Pool that
was fetched directly.  Only VMs/Pools fetched by page would
be refreshed and maintained.  Any others would be removed
from the store.  The behavior caused problem when a user would
enter the app at a view or edit page with a VM or Pool that
would normally be on a fetch page >1.  The VM/Pool would be
fetched directly but then removed from the store at the next
refresh.

Change have been made such that any VM/Pool fetched outside of
the paged fetching will also be refreshed:

 - The single VM fetch REST call can now __follow__ all the data
   paths like the fetch by page does. This reduces the number of
   REST calls required from 6 to 1. It now also supports both
   shallow and deep fetches.

 - Adjusted the Selectors to return all VM or Pool IDs currently held
   in the store

 - Fixed a latent bug in the UPDATE_VMPOOLS_COUNT reducer in
   __src/reducers/vms.js__

 - Removal of 'stale' VMs and Pools from the store was moved to
   the refresh saga so it can be done once AND preserve ALL VMs
   and Pools regardless of how they were fetched.

 - `refreshData` now checks VM/Pool ids refreshed by page/count
   against the set of VM/Pool ids currently in the redux store. If
   there are any ids from the store not already updated, they
   will be fetched individually. No valid VM/Pool will be removed
   from the redux store.

NOTE: This change will cause `refreshData` to singly fetch any VM
or Pool not fetched by the paged/by count functions. That could be
for any reason including, but not limited to, accessing a VM or Pool
that is past the first fetch page via direct URL (the core of
issue #574), deletion of the VM or Pool, permission changes to the
VM or Pool, or if the VM or Pool gets pushed past the last page the
App is currently refreshing.