ux-redesign: refactor getRoutes() route title functions
A route's title function is now passed __vms__ as the second
argument.

With the `ConnectedRouter` in place, passing __vms__ to `getRoutes()`
was rendering the route but not changing the render when __vms__ is
updated. The route doesn't change even if the __vms__ does, so titles
would not get updated.

The only need for __vms__ is to pull a VM's name for the route's
title used by `Breadcrumb`. On a browser page reload for any route
that uses __vms__ (VM/Pool Details and Edit), the Breadcrumb will
render the VM/Pool ID initially. However, it didn't get updated
to the VM/Pool name since getRoutes() was holding on to an old
copy of __vms__. Pushing the __vms__ to the `title` functions ensures
the correct copy of __vms__ is used every time.