Fix the checks for hooking redux devtools extension in dev mode
A missing check for the presence of the redux devtools extension compose
hook function was preventing the app from loading when the extension
was not installed on the browser and the app is running for the webpack
dev server. The check has been added and the app will run from the dev
server without the extension installed and active on the browser.