api: on disk ops, poll attachment for updates
When waiting for an operation to complete, both
the attachment and the disk itself need to be
queried to look for **disk status** and the
attachment **active** flags.

If only the disk itself is polled, the **active** flag
won't get updated along with the **status** indicator
even though it has been updated.