refresh data: update ISOs only on manual refresh
Since the ISO list on an environment is expected to stay relatively
static, loading the list once at login will be sufficient.  If a
user needs to see a newly uploaded ISO, they may either reload the
app (browser reload) or manually click the refresh icon in the
masthead.  Either of those operations will fetch a current list of
ISOs.