add meta tag to keep IE11 off compatibility mode
web-ui works ok on IE11 both in dev and prod mode, but if intranet
settings get turned on and IE go to "compatibility mode", it doesn't
work.  Adding the meta tag to the static app loader page keeps IE
out of that mode.

If a company has "Enterprise Mode" turned on, web-ui will need to
be white listed.  No meta tags in HTML headers can coerce IE11 into
working differently in that scenario.