Verify translated message formats match English
Adds 2 new tests to verify the contents of `translated-messages.json`
against the English source text in `messages.js`.

  1. Verify the translations only have keys that are already defined
     English text keys

  2. For messages that are ICU formatted messages, verify the
     translations have the same count and named arguments as the
     English message.

A failure in test 1 will not prevent the app from functioning.

A failure in test 2 will cause the locale with the failure to
break at run time.  Typically this will only happen if (1) the
translation has a typo in the argument name, or (2) if the argument
name is changed in the English text but not carried forward to the
translations.

NOTE: If a fix the the translations is required to pass the test,
      the same change should be made in Zanata so it is retained
      in the next translation update.