Add explicit tests for each localized message
Upgraded Jest to 24.0.0 to have access to `describe.each`. This allows
each localized key to be its own test. If a key's test fails, we know
exactly which key and why.

Fixed `normalize-messages.js`. It was using an undeclared import and
that import was dropped by the Jest 24.0.0 upgrade. The devDependencies
have been updated so this won't happen again.