Add 'canUserChangeCd' permit check to VM transform
To support a normal user's ability to change a VM's CD, a field
has been added to a VM's transform: `canUserChangeCd`.  This is
the basis for DetailsCard permission check to be able to edit
the VM.

API call `getVmPermissions` was updated to be correct, even if
it won't ever really be used.