`editVm` saga to skip VM update if only CD change is needed
On the `DetailsCard`, it is quite possible the only change that
needs to be committed is to change the CD.  In this case, the
VM edit call can be skipped.  Additionally, when the user does
not have access to edit the VM, but can change the CD, it is
essential to skip the VM update call.

The `editVm` saga now looks at the given `vm` object and if only
fields required to change the CD are present, the VM update is
skipped.  The ability to specify if the change CD should be only
temporary on the running VM or an update to the default CD on the
VM itself was added.