At login, fetch all roles available to the user
By fetching all of the roles, with permits, that are available to a
user at login time, the set of permits available to the user based on
their permission set can be calculated on the front end without having
to fetch the permits directly.

We can't fetch the set of permits directly when requesting an entity
since the permits are on the 3rd tier down from the entity and the
REST 'follow' parameter doesn't traverse that far down. [1]

Now, we can fetch an entities permissions and cross reference each
permission's role with the one in the redux store and build a permit
list from there without having to go back to the API.

[1] - https://bugzilla.redhat.com/show_bug.cgi?id=1639784