Resolve issues with immediate logout when app starts
When the app is starting up, the `SessionActivityTracker` would
immediately load. If the tracker started before the timeout
value is loaded, as part of the login initial data fetch, it was
considering a timeout value of `null` as 0 and immediately logging
out the user.

This has been fixed in two ways:
  1. The `SessionActivityTracker` isn't loaded on the `App` until the
     app is ready (i.e. once enough initial data has been fetched)

  2. `SessionActivityTracker` will not tick down its counter when
     the counter is set to `null` (i.e. the `userSessionTimeoutInterval`
     hasn't been fetched yet)

Fixes: #1024