Allow `FieldValue` component to have undefined children/tooltips
When rendering a VM that is not running, the VM's __host__ is
undefined.  This causes the `FieldValue` proptype checker to throw
an error on the browser console.  The child value is in the JSX code,
but the value is `undefined`.

Updated the component's prop type such that `children` and `tooltips`
must both be defined or must both be undefined.  If they are not
defined, don't bother rendering anything.