Remove logger.js in favor of straight console logging
`logger.js` is structured in such a way that standard console logging
is enhanced and is identical in functionality to using the exported
logger.  To maintain consistency across the code base for logging,
the exported version of logger has been removed and its previous usage
has been replaced by traditional console logging.