Only show file system chart when there is >1 file system
Fixes: #1055

There is a permutation where the standard QEMU guest agent
will not report actual per file system data.  In this case, the
VM's `disks.usage` statistic is an empty array.  This case is
now tested, and if the usage array is empty, the unallocated vs
provisioned donut chart is displayed (as if the guest agent is
not installed).