Add a warning alert to the Disk editor 'bootable' checkbox
On the disk image editor, if the current disk being edited or created
is marked to be bootable, and it is not the current bootable disk,
display a warning message that the bootable flag will be moved.

Don't show the message if the VM doesn't have disks or if none of
the VM's disks have been marked bootable.