Fix DiskImageEditor storage domain list
A regression was introduced in the create VM code such that
the storage domain list on the disk editor/creator modal showed
as blank.  This has been fixed by adding the proper data center
filtering to the list generation.

Now, the storage domain list will only show storage domains available
for use by the logged in user in the VM's cluster's data center.