Added External Service Button.
Added a service button to the VM's action panel.

VM Custom Properties are now fetched by the ovirtapi library.
Updated ovirtapi to store all custom properties.

Custom VM Properties can now be used to define the custom service,
name, and port. These go into a custom property calles external_service.

Updated frontend to parse the external_service property.

Resolved issue with onClick actions not working in card view.

The onClick action from the VmAction was not being passed to the
drop down action list that is shown in the VM Card view. This hadn't
been an issue before because the other actions use a confirmation
dialog, not an onClick action.