Fix VM reducer UPDATE_VMPOOLS_COUNT
There are scenarios when a VM belongs to a Pool (i.e. it has
a pool id associated with it) but the Pool has not yet been
fetched.  The UPDATE_VMPOOLS_COUNT reducer has been updated to
account for this.  If a VM is tagged to a pool that doesn't exist,
the count operation for that pool is skipped.

This leaves a scenario where a VM may momentarily appear as a normal
VM but the get tagged as a Pool VM once the Pool is actually fetched.

The UI impact isn't desirable, but VM Portal will no longer break
when this scenario is encountered.