Simplify UPDATE_VMS reducer handling of 'copySubResources'
The incoming vms are plain JS objects.  It is easier to handle
copying of existing sub-resources from the current copy of the
VM in redux to the incoming VM updates when they're both JS objects.

The reducer converts the merged VM data to ImmutableJS before merging
it to the redux state.