Add Bootable column in the Storage step of a Create VM Wizard
Fixes: https://github.com/oVirt/ovirt-web-ui/issues/1199

Display the column only when adding/editing a disk.
Display the checkbox for an appropriate disk (row) which is being edited.
Render the Checkbox component only when it's possible to set actual disk as bootable.
Add the icon - tooltip in the Bootable column and appropriate row to inform user about:
- why they cannot set any disk as bootable if they cannot (presence of a bootable template disk)
- that only one disk at a time can be set as bootable
- that the bootable flag will be removed from a bootable disk and placed on the actual disk if possible