On logout, set the page to NO_REFRESH and stop the scheduler
When ovirt-web-ui is installed to ovirt-engine, a logout should push
the user to the base ovirt welcome page.  But when running in dev mode
or via container, the logout page is displayed.  In that case, we want
to make sure the page is set to something appropriate and that
background refreshing is no longer done.