Fix create vm storage step moving bootable flag regression
A regression was introduced with making the VM state immutable. The
storage step was directly updating one of the disks to remove the
bootable flag before sending that object back through the `onUpdate`
callback.  With immutable state via immer, the objects are set to read
only and any updates to the object will throw a runtime error typically
including the phrase "read only property".

The fix is simple.  Do not reuse the object.  Now `onUpdate` is called
with a simple object created just to hold the update work.