Use pf3 ExpandCollapse react component in the VM Details card
Fixes: https://github.com/oVirt/ovirt-web-ui/issues/1296

Use the pf3 react component instead of our implementation of the expand/collapse component in the VM Details card.
Also add appropriate css style to the pf component to achieve the expected look (same as before, extra space above
the Advanced Options, proper align of the expand/collapse icon with the rest of the component).