Update parsing of VM statistics (fetched with VM details)
This commit changes the check for the `disks.usage` statistic
to exclude any falsey datum value from further parsing when
processing a VM's statistics.

With changes on the master branch ovirt-engine from at least 4.4.4.2,
VM statistics for file system usage is responding slightly different.
As a result, the VmStatistics transform code was returning an
`undefined` for the `disks.usage` named statistic instead of the
anticipated `null`. Since the code was explicitly checking for a `null`
value, the `undefined` datum was not excluded from further attempted
parsing.  `undefined` can't be further parsed and the errors happen.