Fix create VM from template bootable disk flag check
In the Create VM Storage step, if any of a VM's template defined disks
is marked bootable, then the bootable flag cannot be changed when
adding or editing a new disk.

Previously, the `isBootableDiskTemplate` check in the `Storage`
component worked for a template provisioned new VM with a single
template defined disk, but not for templates that have multiple
defined disks.  The check has been updated to handle multiple template
defined disks, any of which may be bootable.