A temp fix for reading new format of config val MaxNumOfVmCpus
Since the MaxNumOfVmCpus config value was changed to a Map on backend in
patch https://gerrit.ovirt.org/113576, then we need to parse the config
val for supporting that per version.

As a temp & quick solution, this value in just hardcoded in case of
engine >=4.4.5 in this PR

This will be replaced by the actual parsing in a followed patch.