Refactor engine option handling in general and `MaxNumOfVmCpus` in specific
Engine option handling
----------------------

Previously, engine options with multiple version/values incorrectly
used the engine's version to pick the correct value.  The option
version corresponds to the "effective compatibility version".  The
actual version to look at is variable based on what operation is
being down with what information.  The base version values can be
found in (from most general to most specific):
  - DataCenter version
  - Cluster version
  - Template `custom_compatibility_version`
  - VM `custom_compatibility_version`

Engine options with a single 'general' version needed no changes.

The engine options used in web-ui that are version sensitive are for
CPU topology:
  - MaxNumOfVmSockets
  - MaxNumOfCpuPerSocket
  - MaxNumOfThreadsPerCpu
  - MaxNumOfVmCpus

Clusters, Templates, and VM fetch sagas now build a `cpuOptions`
map based on that entity's compatibility version.  If a Template
or VM does not define a `custom_compatibility_version`, its
`cpuOptions` map will be empty.

UI components that need to reference `cpuOptions` look at the
VM/Template and then Cluster.  The rules are slightly different for
create and edit so the lookup is not shared.  In each case, the
function `grabCpuOptions()` handles the `cpuOptions` lookup.

`MaxNumOfVmCpus` handling
-------------------------

Update transform and use of `MaxNumOfVmCpus` to align with ovirt-engine
[patch 113576](https://gerrit.ovirt.org/c/ovirt-engine/+/113576).
The patch converts the `MaxNumOfVmCpus` value from a single number
to a Map of numbers.  `MaxNumOfVmCpus` is now a Map of architecture
type to a number.

Changes made to handle the new value:

  - Parse the value with a new Transform (the value is not JSON)

  - Use the cluster's architecture to select the `MaxNumOfVmCpus`
    based on that architecture

  - If the cluster architecture cannot be mapped, a default value
    is used

For reference on ovirt-engine, this is the Java method that verifies
CPU topology limits using config values looked up by compatibility
version and CPU architecture:
  - backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/validator/VmValidator.java#validateCpuSockets()