Update how template disks are created
Changes have been made to following how Admin Portal works when
creating VMs from a template.  How template disks are created depends
on (1) the Template and VM's optimized for selection, and (2) if a new
storage domain needs to be selected.

The VM's "Optimized For" selection drives the default disk type for
the template disks.  In Admin Portal, selecting **Desktop** resets
the "Storage Allocation" selection to **Thin**.  Selecting **Server**
or **High Performance** resets "Storage Allocation" to **Clone**.

  If VM's optimized for selection is **Desktop**:
    - behavior matches when Admin Portal **Storage Allocation** is **Thin**
    - all of the template disks' disk types are forced to "Thin Provisioned"

    - the VM create is sent:
      - `clone=true` if AT LEAST ONE of the disks has a storage domain
        that differs from the template's disk
      - `clone=false` as long as all of the template disk's storage
        domains remain unchanged

  If VM's optimized for selection is **Server**:
    - behavior matches when Admin Portal's Storage Allocation** is **Clone**
    - the template disks use the disk type as defined by the template

    - the VM create is sent `clone=true`

Fixes: #1371