Update create VM step navigation and disk create checks
Fixes: #1366

Creating a VM from a template with disks defined, and if the
disks need to be cloned to a new storage domain, navigation
buttons needed to be enabled differently.  Now, once on the
storage domain wizard step:

  - The forward button is enabled ONLY when all of the disks are
    correct and no edit is active.

  - The back button is always enabled, including when:

    - an edit is active

    - no storage domain is available for the template disks

    - disk storage domains must be selected, but have not
      yet been selected (i.e. edits are active)

When no storage domains are available, and no disks have been
defined by the selected template, the `EmptyState` place holder
will now show a "cannot create a disk" message.  This message
replaces the "Create Disk" button.