Update `entityPermissionsToUserPermits()` to correctly filter permissions
Fixes: #1411

The assumption that fetching permissions with `Filter: true` will
only return permissions that **apply** to requesting user was wrong.
All permissions that can be **read by** the requesting user are
returned.  Additional filtering needs to be done to make sure the
permissions that can be seen apply to the user (or a group the user
belongs to).

Now `entityPermissionsToUserPermits()` will verify that a permission
applies to the user before adding the permission's permits to the
set of permits for an entity.