Refactor BasicSettings and DetailsCard `grabCpuOptions()`
Now that the saga fetch functions will set `cpuOptions` to `null`
if a VM or Template does not have a custom capability level defined,
the `grabCpuOptions()` function can be simplified.

Each of the 4 entries that are used by the components are still
explicitly retrieved to make it obvious in code where those values
are coming from.