Refactor webpack, babel, jest and polyfills
Webpack changes to dev and prod were a combination of the existing
upgrade work and referencing the configs/setup of a current create
react app (CRA) that was ejected and the configs/setup of
ovirt-engine-ui-extensions.  Being aligned with the CRA allows easier
adoption of their tooling upgrades.  Being aligned with ui-extensions
makes the ovirt project more similar and easier to maintain.

Note: The dev and prod configurations are very close to each other and
in future work common pieces could be refactored out.

Babel changes add a basic set of presets and plugins following
the patterns from CRA's babel-preset-react-app definitions.  The
configs are kept internal to the app to allow finer control over
polyfills and babel plugins.

Jest configs updated to work with the new app based babel preset.

Removed some unused pacakges.