Cross reference user domain groups with ovirt groups
At login, we want to know the uuids of the ovirt added
domain user groups for the logged in user.  Group membership
drives permission look ups.  Since the `/api/users/<userid>/groups`
returns the domain user groups, additional work is required to
figure out which of those groups are ovirt groups.

Now at login, the domain groups are cross referenced with the
ovirt know user groups (i.e. domain groups that have been added
to ovirt in the Administration > Users > Group panels).  This
result of the cross reference is a list of the ovirt uuids of
the user groups the user is a member of.  With the ovirt uuids
the permission calculations used to enable/disable functions of
the app will work correctly.

Resolves: #1438