Update invalid vCpu handling on VM Detail > Details Card
When the vCPU value is out of range, then:
  - Do not calculate the CPU topology on edit
  - Do not calculate the CPU topology on render
  - Limit the maximum input value via constant to 4096

This will allow entering a value that is out of range but then
change the cluster where the value may now be in range.  Arbitrarily
large values are not allowed.  Very large values were causing browsers
to freeze while attempting to calculate valid CPU topologies.