Don't display manual refresh or setting until the app is ready
Just like how the main `PageRouter` is not displayed until the app
is marked as ready to run, do not show the toolbar actions that require
the app to be ready until the app is ready.  The manual refresh action
and the global settings action should not be available until the app
is ready to run.

Note: The app is marked as ready to run as the last part of the login saga.