Refactor all VM and Pool fetching / putting
For both VM and Pool fetching, page fetching, and background
refreshing, split the fetch/transform portions.  This allows the
sagas to focus on the fetching things, making the entities look
correct, and deciding what needs to be done to the store.

The putting is currently all done at the end of each saga.  With
this commit, multiple puts are still done.  The next step is to refactor
the reducers so the sagas can do a single put.  A single put/reducer
is how we make sure the store updates are atomic.

Drop selectVmDetail() and selectPoolDetail() actions/sagas
  - replace `selectVmDetail` action/saga with `getSingleVm()`/`fetchSingleVm()`
  - replace `selectPoolDetail` action/saga with `getSinglePool()`/`fetchSinglePool()`

Refactor vm action sagas
  - check job status to refresh quickly after action is completed
  - move start/stopProgress sagas