vms reducer UPDATE_VMS handles all base updates
All VM and Pool collection operating reducers have been refactored
into a single reducer.  This allows all of the updates, deletes,
and pagination tracking change to be handled as a single atomic update
to the redux store.

Going forward, sagas should only put to a single reducer when saving
data to the store.  Having multiple puts only allows for race conditions
and for the store to get corrupted.