Handle bad vm/pool by count fetch in `refreshListPage`
In the `refershListPage` reducer, if a VM or Pool fetch by count call
fails (for any reason) then the rest of the saga still needs to run.
Look for a `null` response from `fetchVms()` and `fetchPools()` and
use appropriate values in this error case so the rest of the saga
will still function properly.