Display numeric values in CPU Utilization sparkline chart in VM details
Fixes: https://github.com/oVirt/ovirt-web-ui/issues/1504

Display the numeric values in the tooltip for CPU utilization sparkline chart
within VM details dialog.

After upgrade of the @patternfly/react-charts package, 'undefined' was displayed
insted of the number, in the CPU chart. Same for the Memory and Networking
utilization sparkline charts of the VM details. This was due to changes in the
package, Chart react component.

Without this commit, using Chart component and any version of the package
higher than 5.0.0 (incl.) would lead to the original problem.