Update storageDomain transforms and diskFormatToSparse handling
Use `templateDiskFormatToSparse` as a function taking in format,
config items and the template's disk and returning the proper
sparse value for template disk creation.  Using a mapping function
with the extra values as parameters works better than trying to
calculate all permutations at transform time.

Any sagas using this mapping will need to lookup and pass in the
config values as appropriate.