Assign snapshot pending tasks to vm/snapshot
Previously the `pendingTasks` for snapshot add, restore, and remove
were stored globally.  This blocked ALL snapshot actions for ALL VMs
when one task was running.

Snapshot action `pendingTasks` are now tagged to the vm/snapshot the
action is working on.  This will allow snapshot actions to run
independently between VMs.

Fixes: #1559