Allow selection of ISO images from storage domains tagged `canUserUseIsoImages`
Fixes: #1572

A user that has the proper permits (`edit_vm_properties`) to edit a VM's
CD should be allowed to choose any ISO from any Storage Domain in which
they have an `attach_disk_profile` permit.

Add a flag to Storage Domains called `canUserUseIsoImages` to indicate
the current user can see and select and ISO held in that storage domain.
A user is not required to have `canUserUseDomain` on a SD to be able to
use the ISOs in the SD.