Recover after rejected fullscreen request
Fullscreen request can be rejected due to various reasons. In such
case the isFullScreen flag needs to be cleared to synchronize the state
and the displayed UI.

Note that fullscreen request cannot be triggered automatically on page
load. Due to security reasons the API is restricted to user gestures.
This prevents opening the console automatically in full screen when
accessed directly via the link (i.e. by page refresh).