Lock snapshot during restore operation
The snapshot is locked on the backend  during the restore but the UI
needs to wait until the next refresh to receive this status change.
To provide better user experience we can lock the snapshot right
after restore is triggered. The next refresh will confirm or clear
the status (according to restore operation progress).